Modify: disable drag and drop and fix typo
