+++ src/helper/card.js
@@ -46,4 +46,32 @@ function validateSet(set) {
   return isCorrectSet;
 }
 
-export { getAllCardInfo, shuffleCards, validateSet };
+function findValidSet(openedCards) {
+  const correctSet = [];
+
+  for (let i = 0; i < openedCards.length - 2; i++) {
+    if (correctSet.length) {
+      break;
+    }
+
+    for (let j = i + 1; j < openedCards.length - 1; j++) {
+      if (correctSet.length) {
+        break;
+      }
+
+      for (let k = j + 1; k < openedCards.length; k++) {
+        const set = [openedCards[i], openedCards[j], openedCards[k]];
+        const hasCorrectSet = validateSet(set);
+
+        if (hasCorrectSet) {
+          correctSet.push(i, j, k);
+          break;
+        }
+      }
+    }
+  }
+
+  return correctSet;
+}
+
+export { getAllCardInfo, shuffleCards, validateSet, findValidSet };
